+++ lmcache_ascend/v1/storage_backend/__init__.py
@@ -61,7 +61,7 @@ def CreateStorageBackends(
 
     if config.enable_pd:
         # First Party
-        from lmcache_ascend.v1.storage_backend.pd_backend import AscendPDBackend
+        from lmcache_ascend.v1.storage_backend.pd import AscendPDBackend
 
         assert not config.use_layerwise, "Layerwise is not supported for PD"
         storage_backends["PDBackend"] = AscendPDBackend(config, metadata)
