+++ lmcache_ascend/v1/storage_backend/pd/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Ascend PD (Prefill-Decode) backend package."""
+
+from lmcache_ascend.v1.storage_backend.pd.backend import AscendPDBackend
+
+__all__ = ["AscendPDBackend"]
